requirements.txt
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
fastapi
uvicorn
torch
streamlit